Rename feature, entry component in feature and shared ones to details.
