+++ coral/src/app/features/topics/details/components/TopicDetailsResourceTabs.test.tsx
@@ -2,7 +2,7 @@ import { cleanup, screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import { TopicOverviewTabEnum } from "src/app/router_utils";
 import { customRender } from "src/services/test-utils/render-with-wrappers";
-import { TopicOverviewResourcesTabs } from "src/app/features/topics/overview/components/TopicOverviewResourceTabs";
+import { TopicOverviewResourcesTabs } from "src/app/features/topics/details/components/TopicDetailsResourceTabs";
 import { within } from "@testing-library/react/pure";
 
 const mockedNavigate = jest.fn();
@@ -42,7 +42,7 @@ const testMapTabs = [
   },
 ];
 
-describe("TopicOverviewResourceTabs", () => {
+describe("TopicDetailsResourceTabs", () => {
   const user = userEvent.setup();
 
   const testTopicName = "my-nice-topic";
