+++ coral/src/app/features/topics/details/components/TopicDetailsResourceTabs.tsx
